backend/requirements.txt
fastapi>=0.115.0,<1
uvicorn[standard]>=0.32.0,<1
python-multipart>=0.0.9
pydantic>=2.10.0,<3
anthropic>=0.40.0,<1
httpx>=0.28.0,<1
Pillow>=10.0.0
librosa>=0.10.1
numpy>=2.0.0,<3
soundfile>=0.12.1
